refactor(renderer): extract metadata + texture modules from renderer_objects
renderer_objects.cpp was 3646 lines. Move the JSON/QVM metadata loaders (portal/window/AI/deathzone/magic ids) into renderer_objects_metadata.cpp and the texture/DAT mapping, lookup, and application path into renderer_objects_texture.cpp (with its StripTextureFormatSuffix/TextureIdHasAlpha helpers). The shared IsWeaponModel helper and common includes move to renderer_objects_internal.h. Core drops to 2832 lines; ATTA/picking/visual remain for a follow-up. Pure code motion: builds clean and renders textures at runtime (level 2, 93 texture applications, no errors).
