+++ source/renderer/renderer_objects.cpp
@@ -1,316 +1,8 @@
-#include "pch.h"
-#include "renderer_objects.h"
-#include "../config.h"
-#include <iostream>
-#include <fstream>
-#include <filesystem>
-#include <map>
-#include <set>
-#include <glm/gtc/matrix_transform.hpp>
-#include <glm/gtc/type_ptr.hpp>
-#include "logger.h"
-#include "utils.h"
-#include "../level/level_common.h"
-#include "gl_helper.h"
-#include "../parsers/mef_native.h"
-#include "../parsers/qvm_parser.h"
-#include "../parsers/qvm_decompiler.h"
-#include "../parsers/dat_parser.h"
-#include "../parsers/res_compiler.h"
-#include "../parsers/mtp_parser.h"
-#include <sstream>
-
-// Strip pixel-format suffixes that appear in DAT texture IDs but aren't part
-// of the actual .tex filename on disk (e.g. "009_09_1_argb8888" → "009_09_1").
-static std::string StripTextureFormatSuffix(const std::string& texId) {
-    static const char* const kSuffixes[] = {
-        "_argb8888", "_rgb565", "_argb1555", "_argb4444",
-        "_a8r8g8b8", "_r5g6b5", "_a1r5g5b5", "_a4r4g4b4"
-    };
-    for (const char* suf : kSuffixes) {
-        const size_t sufLen = std::strlen(suf);
-        if (texId.size() > sufLen) {
-            // Case-insensitive suffix check
-            std::string tail = texId.substr(texId.size() - sufLen);
-            std::string sufLower(suf);
-            for (auto& c : tail)     c = static_cast<char>(std::tolower(static_cast<unsigned char>(c)));
-            for (auto& c : sufLower) c = static_cast<char>(std::tolower(static_cast<unsigned char>(c)));
-            if (tail == sufLower) {
-                return texId.substr(0, texId.size() - sufLen);
-            }
-        }
-    }
-    return texId;
-}
-
-// True if a DAT texture id is flagged with an alpha-bearing pixel format
-// (e.g. "009_09_1_argb8888"). These textures carry real per-texel transparency
-// (sunglasses lenses, visors) that must be alpha-blended; the opaque pass would
-// otherwise hard-cutout the lens (shader: discard when texColor.a < 0.5) and
-// render it as a distorted black blob. RGB565/_r5g6b5 have no alpha.
-static bool TextureIdHasAlpha(const std::string& texId) {
-    static const char* const kAlphaSuffixes[] = {
-        "_argb8888", "_argb1555", "_argb4444",
-        "_a8r8g8b8", "_a1r5g5b5", "_a4r4g4b4"
-    };
-    for (const char* suf : kAlphaSuffixes) {
-        const size_t sufLen = std::strlen(suf);
-        if (texId.size() > sufLen) {
-            std::string tail = texId.substr(texId.size() - sufLen);
-            for (auto& c : tail) c = static_cast<char>(std::tolower(static_cast<unsigned char>(c)));
-            if (tail == suf) return true;
-        }
-    }
-    return false;
-}
+#include "renderer_objects_internal.h"
 
 
-// ─── EnsurePortalDistancesLoaded ──────────────────────────────────────────────
-void Renderer_Objects::EnsurePortalDistancesLoaded() {
-    if (portal_distances_loaded_) return;
-    portal_distances_loaded_ = true;
-
-    std::string qvmPath = Utils::GetIGIRootPath() + "\\lod.qvm";
-    if (!std::filesystem::exists(qvmPath)) {
-        Logger::Get().Log(LogLevel::WARNING, "[Renderer_Objects] lod.qvm not found at: " + qvmPath);
-        return;
-    }
-
-    QVMFile qvm = QVM_Parse(qvmPath);
-    if (!qvm.valid) {
-        Logger::Get().Log(LogLevel::ERR, "[Renderer_Objects] Failed to parse lod.qvm");
-        return;
-    }
-
-    std::string decompiled = QVM_DecompileToString(qvm);
-    if (decompiled.empty()) return;
-
-    std::stringstream ss(decompiled);
-    std::string line;
-    while (std::getline(ss, line)) {
-        // Look for: Task_New(-1, "ModelLODSettings", "300_01_1", "300_01_1", 70.0, 170.0, 300.0, 350.0, 500.0);
-        if (line.find("Task_New") == std::string::npos) continue;
-        if (line.find("\"ModelLODSettings\"") == std::string::npos) continue;
-
-        size_t start = line.find('(');
-        size_t end = line.rfind(')');
-        if (start == std::string::npos || end == std::string::npos) continue;
-
-        std::string args = line.substr(start + 1, end - start - 1);
-        
-        std::vector<std::string> tokens;
-        std::stringstream argStream(args);
-        std::string token;
-        while (std::getline(argStream, token, ',')) {
-            token.erase(0, token.find_first_not_of(" \t\r\n"));
-            token.erase(token.find_last_not_of(" \t\r\n") + 1);
-            tokens.push_back(token);
-        }
-
-        if (tokens.size() >= 9) {
-            std::string modelId = tokens[2];
-            if (modelId.size() >= 2 && modelId.front() == '"' && modelId.back() == '"') {
-                modelId = modelId.substr(1, modelId.size() - 2);
-            }
-            try {
-                float portalDist = std::stof(tokens[4]);
-                portal_distances_[modelId] = portalDist;
-            } catch (...) {}
-        }
-    }
-    
-    Logger::Get().Log(LogLevel::INFO, "[Renderer_Objects] Loaded " + std::to_string(portal_distances_.size()) + " LOD distances from lod.qvm");
-}
 
 
-// ─── EnsureWindowModelIdsLoaded ───────────────────────────────────────────────
-// Reads IGIModels.json once and collects ModelIds whose ModelName contains
-// "WINDOW" or "GLASS". These models are rendered semi-transparent.
-void Renderer_Objects::EnsureWindowModelIdsLoaded() {
-    if (window_ids_loaded_) return;
-    window_ids_loaded_ = true;
-
-    const std::string jsonPath = Utils::GetExeDirectory() + "\\editor\\tools\\IGIModels.json";
-    if (!std::filesystem::exists(jsonPath)) {
-        Logger::Get().Log(LogLevel::WARNING, "[Renderer_Objects] IGIModels.json not found at: " + jsonPath);
-        return;
-    }
-
-    std::ifstream f(jsonPath);
-    if (!f.is_open()) return;
-
-    std::string content((std::istreambuf_iterator<char>(f)), std::istreambuf_iterator<char>());
-
-    // Manual parse: find every {"ModelName":"...WINDOW.../GLASS...","ModelId":"..."}
-    size_t pos = 0;
-    while (pos < content.size()) {
-        // Find ModelName value
-        size_t nameKey = content.find("\"ModelName\"", pos);
-        if (nameKey == std::string::npos) break;
-        size_t nameStart = content.find('"', nameKey + 11);
-        if (nameStart == std::string::npos) break;
-        ++nameStart;
-        size_t nameEnd = content.find('"', nameStart);
-        if (nameEnd == std::string::npos) break;
-        std::string modelName = content.substr(nameStart, nameEnd - nameStart);
-
-        // Find ModelId value (must appear after the ModelName within same object)
-        size_t idKey = content.find("\"ModelId\"", nameEnd);
-        if (idKey == std::string::npos) break;
-        size_t idStart = content.find('"', idKey + 9);
-        if (idStart == std::string::npos) break;
-        ++idStart;
-        size_t idEnd = content.find('"', idStart);
-        if (idEnd == std::string::npos) break;
-        std::string modelId = content.substr(idStart, idEnd - idStart);
-
-        // Check if this is a window/glass model
-        auto toUpper = [](std::string s) {
-            for (auto& c : s) c = (char)toupper((unsigned char)c);
-            return s;
-        };
-        const std::string upper = toUpper(modelName);
-        if (upper.find("WINDOW") != std::string::npos || upper.find("GLASS") != std::string::npos) {
-            window_model_ids_.insert(modelId);
-        }
-
-        pos = idEnd + 1;
-    }
-
-    // These glass-room shells in level 12 must render opaque (solid walls, not see-through).
-    window_model_ids_.erase("463_03_1");
-    window_model_ids_.erase("463_04_1");
-
-    Logger::Get().Log(LogLevel::INFO,
-        "[Renderer_Objects] Loaded " + std::to_string(window_model_ids_.size()) +
-        " window/glass model IDs from IGIModels.json");
-}
-
-// ─── EnsureAiModelIdsLoaded ───────────────────────────────────────────────────
-void Renderer_Objects::EnsureAiModelIdsLoaded() {
-    if (ai_ids_loaded_) return;
-    ai_ids_loaded_ = true;
-
-    const std::string jsonPath = Utils::GetExeDirectory() + "\\editor\\tools\\IGIModels.json";
-    if (!std::filesystem::exists(jsonPath)) return;
-
-    std::ifstream f(jsonPath);
-    if (!f.is_open()) return;
-
-    std::string content((std::istreambuf_iterator<char>(f)), std::istreambuf_iterator<char>());
-
-    size_t pos = 0;
-    while (pos < content.size()) {
-        size_t nameKey = content.find("\"ModelName\"", pos);
-        if (nameKey == std::string::npos) break;
-        size_t nameStart = content.find('"', nameKey + 11);
-        if (nameStart == std::string::npos) break;
-        ++nameStart;
-        size_t nameEnd = content.find('"', nameStart);
-        if (nameEnd == std::string::npos) break;
-        std::string modelName = content.substr(nameStart, nameEnd - nameStart);
-
-        size_t idKey = content.find("\"ModelId\"", nameEnd);
-        if (idKey == std::string::npos) break;
-        size_t idStart = content.find('"', idKey + 9);
-        if (idStart == std::string::npos) break;
-        ++idStart;
-        size_t idEnd = content.find('"', idStart);
-        if (idEnd == std::string::npos) break;
-        std::string modelId = content.substr(idStart, idEnd - idStart);
-
-        auto toUpper = [](std::string s) {
-            for (auto& c : s) c = (char)toupper((unsigned char)c);
-            return s;
-        };
-        const std::string upper = toUpper(modelName);
-        if (upper.find("AITYPE_") == 0) {
-            ai_model_ids_.insert(modelId);
-        }
-
-        pos = idEnd + 1;
-    }
-
-    // Also manually include specific hardcoded AI types known to have sunglasses but maybe missing prefix
-    ai_model_ids_.insert("009_01_1"); // Jach Priboi
-    ai_model_ids_.insert("014_01_1"); // Mafia Patrol
-    ai_model_ids_.insert("014_02_1"); // Mafia2 Patrol
-
-    Logger::Get().Log(LogLevel::INFO,
-        "[Renderer_Objects] Loaded " + std::to_string(ai_model_ids_.size()) +
-        " AI model IDs from IGIModels.json");
-}
-
-// ─── EnsureDeathZoneIdsLoaded ──────────────────────────────────────────────────
-// Parses magicobj.qvm (via the QVM decompiler) and collects model IDs registered
-// as TASKTYPE_DEATHZONE. These are invisible trigger zones attached to vehicles/
-// trains via ATTA; they must not be rendered as visual sub-meshes.
-void Renderer_Objects::EnsureDeathZoneIdsLoaded() {
-    if (deathzone_ids_loaded_) return;
-    deathzone_ids_loaded_ = true;
-
-    const std::string qvmPath = Utils::GetIGIRootPath() + "\\magicobj\\magicobj.qvm";
-    if (!std::filesystem::exists(qvmPath)) {
-        Logger::Get().Log(LogLevel::WARNING,
-            "[Renderer_Objects] magicobj.qvm not found at: " + qvmPath);
-        return;
-    }
-
-    QVMFile qvm = QVM_Parse(qvmPath);
-    if (!qvm.valid) {
-        Logger::Get().Log(LogLevel::WARNING,
-            "[Renderer_Objects] Failed to parse magicobj.qvm: " + qvm.error);
-        return;
-    }
-
-    const std::string src = QVM_DecompileToString(qvm);
-
-    // Parse lines of the form:
-    //   DefineMagicObj("model_id", "model_id", TASKTYPE_DEATHZONE);
-    std::istringstream ss(src);
-    std::string line;
-    while (std::getline(ss, line)) {
-        if (line.find("TASKTYPE_DEATHZONE") == std::string::npos) continue;
-        // Extract the first quoted string (the model ID)
-        const size_t q1 = line.find('"');
-        if (q1 == std::string::npos) continue;
-        const size_t q2 = line.find('"', q1 + 1);
-        if (q2 == std::string::npos) continue;
-        deathzone_ids_.insert(line.substr(q1 + 1, q2 - q1 - 1));
-    }
-
-    Logger::Get().Log(LogLevel::INFO,
-        "[Renderer_Objects] Loaded " + std::to_string(deathzone_ids_.size()) +
-        " TASKTYPE_DEATHZONE model IDs from magicobj.qvm");
-}
-
-// ─── EnsureMagicObjIdsLoaded ──────────────────────────────────────────────────
-void Renderer_Objects::EnsureMagicObjIdsLoaded() {
-    if (magicobj_ids_loaded_) return;
-    magicobj_ids_loaded_ = true;
-
-    const std::string qvmPath = Utils::GetIGIRootPath() + "\\magicobj\\magicobj.qvm";
-    if (!std::filesystem::exists(qvmPath)) return;
-
-    QVMFile qvm = QVM_Parse(qvmPath);
-    if (!qvm.valid) return;
-
-    const std::string src = QVM_DecompileToString(qvm);
-    std::istringstream ss(src);
-    std::string line;
-    while (std::getline(ss, line)) {
-        if (line.find("DefineMagicObj") == std::string::npos) continue;
-        const size_t q1 = line.find('"');
-        if (q1 == std::string::npos) continue;
-        const size_t q2 = line.find('"', q1 + 1);
-        if (q2 == std::string::npos) continue;
-        magicobj_ids_.insert(line.substr(q1 + 1, q2 - q1 - 1));
-    }
-}
-
-// ─── InitSphereMesh ───────────────────────────────────────────────────────────
-// Builds a solid unit UV sphere using GL_TRIANGLES. Normal == position on a unit
-// sphere, which gives correct Phong shading without extra math.
 void Renderer_Objects::InitSphereMesh() {
     const int LAT = 12;
     const int LON = 24;
@@ -449,19 +141,6 @@ void Renderer_Objects::DrawMagicObjSpheres(const std::vector<LevelObject>& objec
     glUseProgram(0);
 }
 
-static bool IsWeaponModel(const std::string& modelId) {
-    if (modelId.empty()) return false;
-    if (modelId.size() >= 4 && modelId[0] == '1' && 
-        modelId[1] >= '0' && modelId[1] <= '9' && 
-        modelId[2] >= '0' && modelId[2] <= '9' && 
-        modelId[3] == '_') {
-        return true;
-    }
-    if (modelId.rfind("WEAPON_ID_", 0) == 0 || modelId.rfind("AMMO_ID_", 0) == 0) {
-        return true;
-    }
-    return false;
-}
 
 // METAL_DOOR_SLIDE_UP (model 506_xx) appears in levels 12/13/14 as an EditRigidObj
 // carrying a genuine multi-axis Euler tuple. It bypasses the engine's special door
@@ -2606,500 +2285,6 @@ Mesh Renderer_Objects::GetOrLoadMesh(const std::string& modelId, bool isBuilding
 
 }
 
-std::string Renderer_Objects::GetLevelTexturesPath() const {
-    return Utils::GetExeDirectory() + "\\editor\\textures\\level" + std::to_string(current_level_);
-}
-
-std::string Renderer_Objects::GetLevelTextureDatPath() const {
-    // Try local extracted DAT first
-    const std::string localDat = Utils::GetExeDirectory() +
-        "\\editor\\textures\\level" + std::to_string(current_level_) +
-        "\\level" + std::to_string(current_level_) + ".dat";
-    if (std::filesystem::exists(localDat)) return localDat;
-
-    // Fall back to the game's DAT in the IGI root
-    const std::string root = Utils::GetIGIRootPath();
-    return root + "\\missions\\location0\\level" + std::to_string(current_level_) +
-        "\\level" + std::to_string(current_level_) + ".dat";
-}
-
-void Renderer_Objects::LoadDatIntoMap(const std::string& datPath, std::map<std::string, std::vector<std::string>>& outMap) {
-    DATFile dat = DAT_Parse(datPath);
-    if (!dat.valid) {
-        Logger::Get().Log(LogLevel::WARNING, "[TEX Native] DAT_Parse failed for: " + datPath + " Error: " + dat.error);
-        return;
-    }
-
-    for (const auto& m : dat.models) {
-        // We use operator[] instead of emplace so that later definitions
-        // (which might contain the actual textures) overwrite earlier empty/stale definitions.
-        outMap[m.modelName] = m.textures;
-    }
-}
-
-void Renderer_Objects::EnsureTextureMapLoaded() {
-    if (texture_map_level_ == current_level_) return;
-
-    model_texture_map_cache_.clear();
-    texture_map_level_ = current_level_;
-
-    const std::string datPath = GetLevelTextureDatPath();
-    Logger::Get().Log(LogLevel::INFO, "[TEX Native] Loading DAT map from " + datPath);
-    LoadDatIntoMap(datPath, model_texture_map_cache_);
-    Logger::Get().Log(LogLevel::INFO, "[TEX Native] DAT map loaded level=" + std::to_string(current_level_) + " models=" + std::to_string(model_texture_map_cache_.size()));
-}
-
-void Renderer_Objects::EnsureGlobalTextureMapLoaded() const {
-    if (global_texture_map_loaded_) return;
-    global_texture_map_loaded_ = true;
-
-    const std::string igiRoot = Utils::GetIGIRootPath();
-    const std::string exeDir  = Utils::GetExeDirectory();
-
-    for (int lvl = 1; lvl <= 14; ++lvl) {
-        // Try local extracted DAT first, then game path
-        std::string localDat = exeDir + "\\editor\\textures\\level" + std::to_string(lvl) +
-                               "\\level" + std::to_string(lvl) + ".dat";
-        std::string gameDat  = igiRoot + "\\missions\\location0\\level" + std::to_string(lvl) +
-                               "\\level" + std::to_string(lvl) + ".dat";
-
-        std::string datPath = "";
-        if (std::filesystem::exists(localDat)) datPath = localDat;
-        else if (std::filesystem::exists(gameDat)) datPath = gameDat;
-
-        if (!datPath.empty()) {
-            DATFile dat = DAT_Parse(datPath);
-            if (dat.valid) {
-                for (const auto& m : dat.models) {
-                    global_texture_map_[m.modelName] = m.textures;
-                    model_level_map_[m.modelName] = lvl;
-                }
-                for (const auto& t : dat.allTextures) {
-                    texture_level_map_[t] = lvl;
-                }
-            }
-        }
-    }
-
-    // Also scan the common package (weapons, shared models not tied to any level)
-    std::string commonDat = igiRoot + "\\missions\\location0\\common\\location0.dat";
-    if (std::filesystem::exists(commonDat)) {
-        DATFile dat = DAT_Parse(commonDat);
-        if (dat.valid) {
-            for (const auto& m : dat.models) {
-                if (global_texture_map_.find(m.modelName) == global_texture_map_.end())
-                    global_texture_map_[m.modelName] = m.textures;
-                model_level_map_.emplace(m.modelName, 0);
-            }
-            for (const auto& t : dat.allTextures) {
-                if (texture_level_map_.find(t) == texture_level_map_.end())
-                    texture_level_map_[t] = 0;
-            }
-        }
-    }
-
-    Logger::Get().Log(LogLevel::INFO,
-        "[TEX Native] Global DAT map loaded: " + std::to_string(global_texture_map_.size()) + " models across all levels");
-}
-
-std::vector<std::string> Renderer_Objects::GetTextureIdsForModel(const std::string& modelId) {
-    EnsureTextureMapLoaded();
-
-    // 1. Current level exact match
-    {
-        auto it = model_texture_map_cache_.find(modelId);
-        if (it != model_texture_map_cache_.end()) {
-            return it->second;
-        }
-    }
-
-    // 2. Global DAT search across all other levels exact match
-    EnsureGlobalTextureMapLoaded();
-    {
-        auto it = global_texture_map_.find(modelId);
-        if (it != global_texture_map_.end()) {
-            return it->second;
-        }
-    }
-
-    // 3. Variant fallback within current level: NNN_XX_N -> NNN_01_1
-    {
-        size_t p1 = modelId.find('_');
-        if (p1 != std::string::npos && p1 + 4 < modelId.size()) {
-            std::string primary = modelId.substr(0, p1) + "_01_1";
-            if (primary != modelId) {
-                auto it = model_texture_map_cache_.find(primary);
-                if (it != model_texture_map_cache_.end()) {
-                    return it->second;
-                }
-            }
-        }
-    }
-
-    // 4. Variant fallback in global DAT
-    {
-        size_t p1 = modelId.find('_');
-        if (p1 != std::string::npos && p1 + 4 < modelId.size()) {
-            std::string primary = modelId.substr(0, p1) + "_01_1";
-            if (primary != modelId) {
-                auto it = global_texture_map_.find(primary);
-                if (it != global_texture_map_.end()) {
-                    return it->second;
-                }
-            }
-        }
-    }
-
-    // 5. Fallback to just the prefix (e.g. NNN_XX_X -> NNN)
-    // AI models like "004_02_1" only have "004" mapped in the global DAT.
-    {
-        size_t p1 = modelId.find('_');
-        if (p1 != std::string::npos) {
-            std::string prefix = modelId.substr(0, p1);
-            if (prefix != modelId) {
-                auto it = model_texture_map_cache_.find(prefix);
-                if (it != model_texture_map_cache_.end()) return it->second;
-                
-                auto git = global_texture_map_.find(prefix);
-                if (git != global_texture_map_.end()) return git->second;
-            }
-        }
-    }
-
-    Logger::Get().Log(LogLevel::DEBUG, "[TEX Native] DAT miss (all sources) for modelId=" + modelId);
-    return {};
-}
-
-std::string Renderer_Objects::FindTextureFile(const std::string& textureId) const {
-    // Helper: search one directory for the texture file
-    auto searchDir = [&](const std::filesystem::path& texturesPath) -> std::string {
-        if (!std::filesystem::exists(texturesPath)) return "";
-
-        const std::filesystem::path exactPath = texturesPath / (textureId + ".tex");
-        if (std::filesystem::exists(exactPath)) return exactPath.string();
-
-        for (const auto& entry : std::filesystem::directory_iterator(texturesPath)) {
-            if (!entry.is_regular_file()) continue;
-            if (entry.path().extension() != ".tex") continue;
-            const std::string stem = entry.path().stem().string();
-            if (stem == textureId ||
-                stem.find(textureId) != std::string::npos ||
-                textureId.find(stem) != std::string::npos) {
-                return entry.path().string();
-            }
-        }
-        return "";
-    };
-
-    // 1. Search local extracted textures directory first
-    std::string result = searchDir(std::filesystem::path(GetLevelTexturesPath()));
-    if (!result.empty()) return result;
-
-    // 2. Fall back to the game's own texture directory for this level
-    const std::string igiTexDir = Utils::GetIGIRootPath() +
-        "\\missions\\location0\\level" + std::to_string(current_level_) + "\\textures";
-    result = searchDir(std::filesystem::path(igiTexDir));
-    if (!result.empty()) {
-        Logger::Get().Log(LogLevel::DEBUG,
-            "[TEX Native] Found texture in IGI root level textures: " + result);
-        return result;
-    }
-
-    // 3. Extracted common location0 textures (from location0.res)
-    const std::string commonLocalTexDir = Utils::GetExeDirectory() + "\\editor\\textures\\common";
-    result = searchDir(std::filesystem::path(commonLocalTexDir));
-    if (!result.empty()) {
-        Logger::Get().Log(LogLevel::DEBUG,
-            "[TEX Native] Found texture in extracted common folder: " + result);
-        return result;
-    }
-
-    // 4. Game common/shared textures folder (shared across all levels)
-    const std::string igiCommonTexDir = Utils::GetIGIRootPath() + "\\textures";
-    result = searchDir(std::filesystem::path(igiCommonTexDir));
-    if (!result.empty()) {
-        Logger::Get().Log(LogLevel::DEBUG,
-            "[TEX Native] Found texture in IGI common folder: " + result);
-        return result;
-    }
-
-    // 5. Game location0 common folder (shared across all location0 levels)
-    const std::string igiLoc0CommonTexDir = Utils::GetIGIRootPath() + "\\missions\\location0\\common\\textures";
-    result = searchDir(std::filesystem::path(igiLoc0CommonTexDir));
-    if (!result.empty()) {
-        Logger::Get().Log(LogLevel::DEBUG,
-            "[TEX Native] Found texture in IGI location0 common folder: " + result);
-        return result;
-    }
-
-    // Lazy On-Demand Extraction for cross-level texture references
-    EnsureGlobalTextureMapLoaded();
-    auto tit = texture_level_map_.find(textureId);
-    if (tit != texture_level_map_.end()) {
-        int targetLvl = tit->second;
-        if (targetLvl != current_level_) {
-            Logger::Get().Log(LogLevel::INFO,
-                "[TEX Native] Lazy extracting textures/models on-demand for level " + std::to_string(targetLvl) +
-                " to resolve cross-level texture: " + textureId);
-            AssetExtractor::EnsureLevelAssets(targetLvl, Utils::GetIGIRootPath(), Utils::GetExeDirectory());
-        }
-    }
-
-    // 6. Search all other levels' extracted and game texture directories.
-    //    Textures like "006_07_1" live in level 6's folder but are referenced by
-    //    models appearing in other levels (e.g. 003_02_1 in level 7).
-    const std::string exeDir2  = Utils::GetExeDirectory();
-    const std::string igiRoot2 = Utils::GetIGIRootPath();
-    for (int lvl = 1; lvl <= 14; ++lvl) {
-        if (lvl == current_level_) continue;
-        result = searchDir(exeDir2 + "\\editor\\textures\\level" + std::to_string(lvl));
-        if (!result.empty()) return result;
-        result = searchDir(igiRoot2 + "\\missions\\location0\\level" + std::to_string(lvl) + "\\textures");
-        if (!result.empty()) return result;
-    }
-
-    Logger::Get().Log(LogLevel::ERR,
-        "[TEX Native] Texture NOT FOUND: '" + textureId +
-        "' — searched level " + std::to_string(current_level_) + ", common, and all fallback paths.");
-    return "";
-}
-
-GLuint Renderer_Objects::GetOrLoadTexture(const std::string& textureId) {
-    if (textureId.empty()) {
-        return 0;
-    }
-
-    // Try the EXACT texture id first. The editor extracts format-suffixed files
-    // verbatim (e.g. "405_01_1_argb8888.tex" is a distinct ARGB8888/alpha file that
-    // sits alongside the opaque RGB565 "405_01_1.tex"). Loading the stripped name
-    // first would grab the opaque sibling and render glass/lattice fully opaque.
-    // Only fall back to the stripped id when no exact-suffixed file exists on disk.
-    std::string texturePath = FindTextureFile(textureId);
-    const std::string strippedId = StripTextureFormatSuffix(textureId);
-    if (texturePath.empty() && strippedId != textureId) {
-        texturePath = FindTextureFile(strippedId);
-    }
-    if (texturePath.empty()) {
-        Logger::Get().Log(LogLevel::WARNING, "[TEX Native] Texture search FAILED for ID: " + textureId +
-            (strippedId != textureId ? " (stripped: " + strippedId + ")" : ""));
-        return 0;
-    }
-
-    const std::string cacheKey = std::to_string(current_level_) + ":" + texturePath;
-    auto it = texture_cache_.find(cacheKey);
-    if (it != texture_cache_.end()) {
-        Logger::Get().Log(LogLevel::DEBUG, "[TEX Native] Cache hit textureId=" + textureId + " path=" + texturePath);
-        return it->second;
-    }
-
-    pics_s pics{};
-    if (!Tex_Load(texturePath.c_str(), pics) || !pics.pics_ || pics.num_pic_ <= 0) {
-        Logger::Get().Log(LogLevel::ERR, "[TEX Native] Failed to decode TEX file: " + texturePath);
-        Pic_FreePics(pics);
-        return 0;
-    }
-
-    const pic_s* pic = pics.pics_;
-    Logger::Get().Log(
-        LogLevel::INFO,
-        "[TEX Native] Loaded textureId=" + textureId +
-        " " + std::to_string(pic->width_) + "x" + std::to_string(pic->height_) +
-        " frames=" + std::to_string(pics.num_pic_) +
-        " path=" + texturePath);
-
-    const GLuint texture = GL_RegisterTexture(pic, GL_REPEAT, GL_LINEAR_MIPMAP_LINEAR, GL_LINEAR, true);
-    texture_cache_[cacheKey] = texture;
-    Pic_FreePics(pics);
-
-    Logger::Get().Log(
-        LogLevel::DEBUG,
-        "[TEX Native] Uploaded textureId=" + textureId +
-        " glId=" + std::to_string(texture));
-    return texture;
-}
-
-void Renderer_Objects::ApplyTexturesToMesh(Mesh& mesh, const std::string& modelId, const std::string& parentModelId) {
-    std::vector<std::string> textureIds = GetTextureIdsForModel(modelId);
-    bool isDatMiss = textureIds.empty();
-
-    // When a sub-model has no DAT entry, IGI ATTA sub-models commonly share material slots with their
-    // parent building, so inherit the parent's texture list in that case.
-    if (!parentModelId.empty() && isDatMiss) {
-        const std::vector<std::string> parentIds = GetTextureIdsForModel(parentModelId);
-        const bool parentHasRealEntry = !parentIds.empty();
-        if (parentHasRealEntry) {
-            Logger::Get().Log(LogLevel::INFO,
-                "[TEX Native] Sub-model '" + modelId + "' has no DAT entry; inheriting " +
-                std::to_string(parentIds.size()) + " texture(s) from parent '" + parentModelId + "'");
-            textureIds = parentIds;
-            isDatMiss = false;
-        }
-    }
-
-    // Conversely, some vehicles or magic models (_01_1) have no textures in DAT, but their interior 
-    // attachments (_02_1) DO have textures. If we still have a DAT miss, inherit from the typical child.
-    if (isDatMiss) {
-        size_t p1 = modelId.find("_01_1");
-        if (p1 != std::string::npos) {
-            std::string childId = modelId.substr(0, p1) + "_02_1";
-            std::vector<std::string> childTexIds = GetTextureIdsForModel(childId);
-            if (!childTexIds.empty()) {
-                Logger::Get().Log(LogLevel::INFO,
-                    "[TEX Native] Model '" + modelId + "' has no DAT entry; inheriting " +
-                    std::to_string(childTexIds.size()) + " texture(s) from child '" + childId + "'");
-                textureIds = childTexIds;
-                isDatMiss = false;
-            }
-        }
-    }
-
-    // If it's still a DAT miss (not explicitly in dat), fallback to self name first.
-    if (isDatMiss) {
-        textureIds = { modelId };
-    }
-
-    if (textureIds.empty()) {
-        Logger::Get().Log(LogLevel::INFO, "[TEX Native] No textures mapped for modelId=" + modelId);
-        return;
-    }
-
-    // For AI character models (sunglasses-bearing), load the STRIPPED texture id
-    // so "009_09_1_argb8888" → "009_09_1.tex" (opaque body texture → solid black
-    // glasses). The _argb8888 variant is a tiny 32×32 semi-transparent overlay
-    // used by the game engine; loading it in the editor gives barely-visible lenses.
-    // Non-AI models (guard tower, lattice) need the exact _argb8888.tex for
-    // transparent rendering and keep the "exact first" preference.
-    EnsureAiModelIdsLoaded();
-    const bool isAiModel = ai_model_ids_.count(modelId) > 0;
-
-    std::vector<GLuint> textures;
-    textures.reserve(textureIds.size());
-    for (const auto& textureId : textureIds) {
-        const std::string loadId = (isAiModel && TextureIdHasAlpha(textureId))
-            ? StripTextureFormatSuffix(textureId)
-            : textureId;
-        textures.push_back(GetOrLoadTexture(loadId));
-    }
-
-    if (!mesh.subMeshes.empty()) {
-        size_t assigned = 0;
-
-        // Find the best valid (non-zero) texture to use as fallback for submeshes
-        // that fall outside the DAT texture list range.
-        GLuint fallbackTexture = 0;
-        for (const GLuint t : textures) {
-            if (t != 0) { fallbackTexture = t; }
-        }
-
-        for (size_t i = 0; i < mesh.subMeshes.size(); ++i) {
-            GLuint texture = 0;
-            const int matSlot = mesh.subMeshes[i].materialSlot;
-            std::string resolvedTexId; // tracks the actual texture id string for ARGB detection
-
-            if (textures.size() == 1) {
-                texture = textures[0];
-                resolvedTexId = textureIds.empty() ? "" : textureIds[0];
-            } else if (matSlot >= 0 && static_cast<size_t>(matSlot) < textures.size()) {
-                texture = textures[matSlot];
-                resolvedTexId = (static_cast<size_t>(matSlot) < textureIds.size()) ? textureIds[matSlot] : "";
-            } else if (matSlot > 0 && !textures.empty()) {
-                EnsureAiModelIdsLoaded();
-                if (ai_model_ids_.count(modelId) > 0) {
-                    // materialSlot is out of range — sunglasses-bearing characters (009_01_1,
-                    // 014_01_1, 014_02_1) where the MEF block references a slot beyond the
-                    // local DAT texture list. Resolve via global DAT.
-                    bool resolved = false;
-                    EnsureGlobalTextureMapLoaded();
-                    {
-                        auto git = global_texture_map_.find(modelId);
-                        if (git != global_texture_map_.end() &&
-                            static_cast<size_t>(matSlot) < git->second.size()) {
-                            const std::string& globalTexId = git->second[static_cast<size_t>(matSlot)];
-                            GLuint globalTex = GetOrLoadTexture(globalTexId);
-                            if (globalTex) {
-                                texture = globalTex;
-                                resolvedTexId = globalTexId; // preserve for ARGB detection
-                                resolved = true;
-                                Logger::Get().Log(LogLevel::INFO,
-                                    "[TEX Native] AI materialSlot out of range resolved via global DAT for modelId=" + modelId +
-                                    " submeshIndex=" + std::to_string(i) +
-                                    " materialSlot=" + std::to_string(matSlot) +
-                                    " globalTexId=" + globalTexId);
-                            }
-                        }
-                    }
-                    if (!resolved) {
-                        texture = textures.back();
-                        resolvedTexId = textureIds.empty() ? "" : textureIds.back();
-                        Logger::Get().Log(LogLevel::WARNING,
-                            "[TEX Native] AI materialSlot out of range, using last texture for modelId=" + modelId +
-                            " submeshIndex=" + std::to_string(i) +
-                            " materialSlot=" + std::to_string(matSlot) +
-                            " textureCount=" + std::to_string(textures.size()));
-                    }
-                } else {
-                    size_t wrapped = static_cast<size_t>(matSlot) % textures.size();
-                    texture = textures[wrapped];
-                    resolvedTexId = (wrapped < textureIds.size()) ? textureIds[wrapped] : "";
-                    Logger::Get().Log(LogLevel::WARNING,
-                        "[TEX Native] materialSlot out of range, wrapping for modelId=" + modelId +
-                        " submeshIndex=" + std::to_string(i) +
-                        " materialSlot=" + std::to_string(matSlot) +
-                        " textureCount=" + std::to_string(textures.size()) +
-                        " wrappedSlot=" + std::to_string(wrapped));
-                }
-            } else if (i < textures.size()) {
-                texture = textures[i];
-                resolvedTexId = (i < textureIds.size()) ? textureIds[i] : "";
-            } else {
-                texture = fallbackTexture;
-            }
-
-            mesh.subMeshes[i].textureID = texture;
-            if (texture) {
-                ++assigned;
-                // ARGB-format textures carry real per-texel alpha (sunglasses lenses,
-                // guard tower lattice, wire fences, etc.). Use resolvedTexId which
-                // captures the actual texture id in ALL code paths including out-of-range
-                // ARGB textures on non-AI models (lattice, fences, decals) need alpha
-                // blending via the transparent pass. AI model ARGB sub-meshes (sunglasses)
-                // must render OPAQUE with the cutout shader so the black lens pixels show
-                // as solid black — blending them makes them see-through.
-                if (TextureIdHasAlpha(resolvedTexId) && ai_model_ids_.count(modelId) == 0) {
-                    mesh.subMeshes[i].alphaMode = 2;
-                    mesh.subMeshes[i].baseColorFactor.a = 0.95f;
-                }
-            }
-        }
-
-        if (!mesh.subMeshes.empty()) {
-            mesh.textureID = mesh.subMeshes.front().textureID;
-        }
-
-        Logger::Get().Log(
-            LogLevel::INFO,
-            "[TEX Native] Applied textures to modelId=" + modelId +
-            " subMeshes=" + std::to_string(mesh.subMeshes.size()) +
-            " datTextures=" + std::to_string(textureIds.size()) +
-            " assigned=" + std::to_string(assigned));
-
-        if (textureIds.size() != mesh.subMeshes.size()) {
-            Logger::Get().Log(
-                LogLevel::WARNING,
-                "[TEX Native] WARNING: Texture/submesh count mismatch for modelId=" + modelId +
-                " subMeshes=" + std::to_string(mesh.subMeshes.size()) +
-                " datTextures=" + std::to_string(textureIds.size()));
-        }
-        return;
-    }
-
-    mesh.textureID = textures.front();
-    Logger::Get().Log(
-        LogLevel::INFO,
-        "[TEX Native] Applied legacy single texture to modelId=" + modelId +
-        " textureId=" + textureIds.front() +
-        " glId=" + std::to_string(mesh.textureID));
-}
 
 Mesh Renderer_Objects::CreateTextMesh(const std::string& text) {
     Mesh m;
@@ -3640,6 +2825,7 @@ void Renderer_Objects::DrawAttachmentsForSpline(
     glDisable(GL_POLYGON_OFFSET_FILL);
     glUseProgram(0);
 }
+
